Server functions, Database functions, Server, and Client Update

+++ database/database.py
@@ -56,7 +56,23 @@ class Camelot_Database():
 
         # If no errors occured, create the account
         cur.execute('''INSERT INTO "USER" VALUES ('{}', '{}')'''.format(username, password))
+        conn.commit()
+
+        # And then add the default channels to the user's channels
+        cur.execute('''
+        SELECT channelid
+        FROM "CHANNEL"
+        WHERE admin IS NULL
+        ''')
+        rows = cur.fetchall()
+
+        for channel in rows:
+            self.add_channels_to_user_info(username, channel)
+
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "success": "Successfully created {}'s account.".format(username)
+        }, indent=4)
 
     ## Validates the username & password are of the correct length
     #
@@ -147,11 +163,25 @@ class Camelot_Database():
         cur = conn.cursor()
 
         for channel in channels:
-            # TODO ZW 3-20: Need to add an error handler for if the user has already joined
-            # a channel, and is trying to join it again.
+            cur.execute('''
+            SELECT userid
+            FROM "CHANNELS_JOINED"
+            WHERE channelid='{}' AND userid='{}'
+            '''.format(channel, username))
+
+            if cur.rowcount == 1:
+                return json.dumps({
+                    "error": "The user has already joined one or more of the channels they were trying to join again."
+                })
+
+        for channel in channels:
             cur.execute('''INSERT INTO "CHANNELS_JOINED" VALUES ('{}', '{}')'''.format(username, channel))
 
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "channels_joined": channels,
+            "user": "{}".format(username)
+        }, indent=4)
 
     ## Creates a channel in the database
     #
@@ -163,6 +193,12 @@ class Camelot_Database():
         conn = self.make_connection()
         cur = conn.cursor()
 
+        # Checks if the channel already exists
+        error = self.check_channel_not_in_database(channel_name)
+        if error:
+            return error
+
+        # Checks to make sure the channel is of the correct length
         if len(channel_name) > 40 or len(channel_name) < 1:
             self.commit_and_close_connection(conn)
             return json.dumps({
@@ -174,8 +210,14 @@ class Camelot_Database():
             cur.execute('''INSERT INTO "CHANNEL" VALUES ('{}', '{}')'''.format(channel_name, admin))
         else:
             cur.execute('''INSERT INTO "CHANNEL" VALUES ('{}', NULL)'''.format(channel_name))
-        self.commit_and_close_connection(conn)
 
+        self.commit_and_close_connection(conn)
+        return json.dumps({
+            "channel_created": {
+                "channel": channel_name,
+                "message": "A new channel has been created: '{}'.".format(channel_name)
+            }
+        }, indent=4)
 
     ## Removes a channel from the database
     #
@@ -212,6 +254,12 @@ class Camelot_Database():
         '''.format(channel_name))
 
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "channel_deleted": {
+                "channel": channel_name,
+                "message": "The channel `{}` has been deleted.".format(channel_name)
+            }
+        }, indent=4)
 
     ## Removes a user from the database
     #
@@ -225,9 +273,11 @@ class Camelot_Database():
 
         # Check for username and password are in database
         error = self.check_username_password_in_database(username, password)
-
         if error:
             return error
+
+        # Get the channels created by the user that will be deleted
+        channels = self.get_channels_user_has_created(username)
 
         # If no errors occur, delete the account
         cur.execute('''
@@ -236,6 +286,12 @@ class Camelot_Database():
         '''.format(username))
 
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "account_deleted": {
+                "username": username,
+                "channels_being_deleted": channels
+            }
+        }, indent=4)
 
     ## Gets all over the users in a specified channel
     #
@@ -277,18 +333,32 @@ class Camelot_Database():
     ## Makes the user leave the specified channel
     #
     #  @param self The object pointer
-    #  @param channel_name The channel specified that the user wants to leave_channel
+    #  @param channel_name The channel specified that the user wants to leave
     #  @param user The user who is wanting to leave a channel
     def leave_channel(self, channel_name, user):
         conn = self.make_connection()
         cur = conn.cursor()
 
+        # Checks if the channel exists in the database
+        error = self.check_channel_in_database(channel_name)
+        if error:
+            self.commit_and_close_connection(conn)
+            return error
+
+        # If the channel does exist, remove the user from the channel
         cur.execute('''
         DELETE FROM "CHANNELS_JOINED"
         WHERE channelid='{}' AND userid='{}'
         '''.format(channel_name, user))
 
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "leave_channel":{
+                "channel": channel_name,
+                "user": user,
+                "message": "{} has left the channel.".format(user)
+             }
+        }, indent=4)
 
     ## Allows the user to change their password
     #
@@ -321,8 +391,34 @@ class Camelot_Database():
         '''.format(new_password, username))
 
         self.commit_and_close_connection(conn)
+        return json.dumps({
+            "success": "Successfully changed {}'s password.".format(username)
+        }, indent=4)
 
-    ## Checks if the channel exists in the database
+    ## Checks if the channel DOES NOT exist in the database
+    #
+    #  @param self The object pointer
+    #  @param channel_name The channel specified to check if it exists in the database
+    def check_channel_not_in_database(self, channel_name):
+        conn = self.make_connection()
+        cur = conn.cursor()
+
+        # Checks if the channel exists in the database
+        cur.execute('''
+        SELECT channelid
+        FROM "CHANNEL"
+        WHERE channelid='{}'
+        '''.format(channel_name))
+
+        if cur.rowcount == 1:
+            self.commit_and_close_connection(conn)
+            return json.dumps({
+                "error": "The specified channel already exists in the database."
+            }, indent=4)
+
+        self.commit_and_close_connection(conn)
+
+    ## Checks if the channel DOES exist in the database
     #
     #  @param self The object pointer
     #  @param channel_name The channel specified to check if it exists in the database
@@ -344,6 +440,83 @@ class Camelot_Database():
             }, indent=4)
 
         self.commit_and_close_connection(conn)
+
+    def check_username_in_channel(self, username, channel_name):
+        conn = self.make_connection()
+        cur = conn.cursor()
+
+        # Checks if user is in specified channel
+        cur.execute('''
+        SELECT userid
+        FROM "CHANNELS_JOINED"
+        WHERE userid='{}' AND channelid='{}'
+        '''.format(username, channel_name))
+
+        if cur.rowcount != 1:
+            self.commit_and_close_connection(conn)
+            return json.dumps({
+                "error": "The user is trying to send a message to a channel they haven't joined yet."
+            }, indent=4)
+
+        self.commit_and_close_connection(conn)
+
+    def new_message(self, username, channel_name):
+        conn = self.make_connection()
+        cur = conn.cursor()
+
+        # Check if the channel exists
+        error = self.check_channel_in_database(channel_name)
+        if error:
+            self.commit_and_close_connection(conn)
+            return error
+
+        # Check if the user is in the specified channel
+        error = self.check_username_in_channel(username, channel_name)
+        if error:
+            self.commit_and_close_connection(conn)
+            return error
+
+        self.commit_and_close_connection(conn)
+
+    # Gets the channels that the user is a part of
+    def get_channels_for_user(self, username):
+        conn = self.make_connection()
+        cur = conn.cursor()
+
+        # Checks if user is in specified channel
+        cur.execute('''
+        SELECT channelid
+        FROM "CHANNELS_JOINED"
+        WHERE userid='{}'
+        '''.format(username))
+        rows = cur.fetchall()
+
+        channels = []
+        for channel in rows:
+            channels.append(channel[0])
+
+        self.commit_and_close_connection(conn)
+        return json.dumps({
+            "channels": channels
+        }, indent=4)
+
+    def get_channels_user_has_created(self, username):
+        conn = self.make_connection()
+        cur = conn.cursor()
+
+        cur.execute('''
+        SELECT channelid
+        FROM "CHANNEL"
+        WHERE admin='{}'
+        '''.format(username))
+        rows = cur.fetchall()
+
+        channels = []
+        for channel in rows:
+            channels.append(channel[0])
+
+        self.commit_and_close_connection(conn)
+        return channels
 
     ## Readies initial data for database
     #
